[MIG] allow to inject fiscal fields in tree
before v14, you could have groups in tree elements. Not anymore in v14.
So we use now use invisible <control> elements as fiscal placeholders in tree
views. And we adapted the view injection to allow to taking back the
tree parent node of these placeholders to append the fiscal fields
directly there.